refactor(test): split tests
move test.js to tests folder
split test.js to smaller test files
